Hub Structure LODs and Tutorial Pivot Adjustments
Fixed topology for the hub structure, put a hole through it, added an elevator asset, made 4 LODs out of it, and imported them into Unity. These changes are in the scene, Hub Structure Collisions. I also moved the center pivots of the movable assets in the Tutorial. These changes are in the scene, Starting Area Testing.