Character Astrolabe Using Animation and Improved Walk/Run
Imported an animation for grabbing the astrolabe out of yourback for use which can be played in reverse to put it back. This is the first one to include extra geometry (astrolabe) so idk how blending will work with it but I was told it should work and that I wouldn't need to make a separate run/walk/ idle, etc. I also imported a pose for holding the astrolabe and improved versions of the walk and run animations.